Move the JSON stringifier to a separate file
Set the `highWaterMark` of the stringifier to our maximum backlog.
Don't `parseInt` options which are supposed to be integers in the first
place.
